+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 * next
   * TDEPS-219 Fixed `-X:deps find-versions` doesn't find Maven versions that start with a letter
+  * Add :n option to `-X:deps find-versions` and default to 8
   * Fix bad invocation of deps/prep-libs! in clojure.tools.cli.api/prep
 * 0.14.1222 on Aug 5, 2022
   * TDEPS-228 Add support for auto inferred Sourcehut git urls